requirements.txt
# ── Web framework ──────────────────────────────────────────────────────────────
fastapi>=0.111.0
uvicorn[standard]>=0.30.0
aiosqlite>=0.20.0
websockets>=12.0
python-multipart>=0.0.9
pydantic>=2.7.0
pydantic-settings>=2.3.0
httpx>=0.27.0

# ── SSH / remote server ────────────────────────────────────────────────────────
paramiko>=3.4.0
cryptography>=42.0.0

# ── Scraping ───────────────────────────────────────────────────────────────────
yt-dlp>=2024.5.27

# ── Audio processing ───────────────────────────────────────────────────────────
soundfile>=0.12.1
noisereduce>=3.0.2
pyloudnorm>=0.1.1
# silero-vad installed separately: pip install silero-vad
# (version depends on your torch version)

# ── Transcription ──────────────────────────────────────────────────────────────
faster-whisper>=1.0.3

# ── Speaker diarization ────────────────────────────────────────────────────────
pyannote.audio>=3.3.1

# ── TTS / training ─────────────────────────────────────────────────────────────
# TTS (Coqui) and peft must be installed AFTER torch.
# See install notes below.
TTS>=0.22.0
peft>=0.11.1
tensorboard>=2.17.0

# ── NOTE: PyTorch ──────────────────────────────────────────────────────────────
# Install torch/torchaudio BEFORE this file to match your CUDA version.
# Example for CUDA 12.1:
#   pip install torch torchaudio --index-url https://download.pytorch.org/whl/cu121
# Example for CUDA 11.8:
#   pip install torch torchaudio --index-url https://download.pytorch.org/whl/cu118
# Then run: pip install -r requirements.txt
